Refactor resolution shader input to resolution provider
